requirements.txt
# Core MCP dependencies
fastmcp>=1.9.1

# LangChain and LangGraph
langchain-core>=0.3.0
langchain-google-genai>=2.0.0
langgraph>=0.2.0
langchain-mcp-adapters>=0.1.0

# Utilities
python-dotenv>=1.0.0
httpx>=0.27.0
fastapi>=0.115.0
uvicorn>=0.30.0

# Development and testing
pytest>=8.0.0
pytest-asyncio>=0.24.0
